DEV-4 moved components connected to Redux to the containers folder
