+++ client/src/containers/TripView/index.js
